% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PriorityQueueWrapper}
\alias{PriorityQueueWrapper}
\title{PriorityQueueWrapper class}
\description{
This class provides an interface to a priority queue in C++ using Rcpp.
It allows you to push, pop, and check the state of the priority queue.
}
