% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{QueueWrapper}
\alias{QueueWrapper}
\title{QueueWrapper class}
\description{
This class provides an interface to a generic queue in C++ using Rcpp.
It allows you to enqueue, dequeue, and check the state of the queue for any R-compatible type.
}
