% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{timesTwo}
\alias{timesTwo}
\title{Multiply a number by two}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{A single integer}
}
\value{
The input multiplied by two
}
\description{
Multiply a number by two
}
\examples{
timesTwo(5)
}
